message_scroll: Extract module message_feed_top_notices.
