+++ tools/test-js-with-node
@@ -110,6 +110,7 @@ EXEMPT_FILES = make_set(
         "web/src/message_edit_history.js",
         "web/src/message_events.js",
         "web/src/message_feed_loading.ts",
+        "web/src/message_feed_top_notices.js",
         "web/src/message_fetch.js",
         "web/src/message_list.js",
         "web/src/message_list_data.js",
